+++ examples/omni_multi/data_classes/_core/__init__.py
@@ -1,5 +1,6 @@
 from omni_multi.data_classes._core.constants import *  # noqa
 from omni_multi.data_classes._core.base import *  # noqa
 from omni_multi.data_classes._core.cdf_external import *  # noqa
+from omni_multi.data_classes._core.datapoints_api import *  # noqa
 from omni_multi.data_classes._core.helpers import *  # noqa
 from omni_multi.data_classes._core.query import *  # noqa