+++ examples/wind_turbine/data_classes/_core/__init__.py
@@ -1,5 +1,6 @@
 from wind_turbine.data_classes._core.constants import *  # noqa
 from wind_turbine.data_classes._core.base import *  # noqa
 from wind_turbine.data_classes._core.cdf_external import *  # noqa
+from wind_turbine.data_classes._core.datapoints_api import *  # noqa
 from wind_turbine.data_classes._core.helpers import *  # noqa
 from wind_turbine.data_classes._core.query import *  # noqa