+++ tests/test_integration/test_sdks/test_scenario_instance_sdk/conftest.py
@@ -1,7 +1,0 @@
-import pytest
-from scenario_instance.client import ScenarioInstanceClient
-
-
-@pytest.fixture()
-def client(client_config) -> ScenarioInstanceClient:
-    return ScenarioInstanceClient.azure_project(**client_config)