+++ tests/test_unit/conftest.py
@@ -6,7 +6,7 @@ from cognite.client.testing import monkeypatch_cognite_client
 from cognite.pygen._core.generators import MultiAPIGenerator, SDKGenerator
 from cognite.pygen._generator import CodeFormatter
 from cognite.pygen.config import PygenConfig
-from tests.constants import CORE_SDK, OMNI_SDK, OMNI_SUB_SDK
+from tests.constants import CORE_SDK, OMNI_SDK, OMNI_SUB_SDK, WIND_TURBINE
 
 
 @pytest.fixture(scope="session")
@@ -73,4 +73,14 @@ def core_multi_api_generator(core_data_model: dm.DataModel[dm.View]) -> MultiAPI
         CORE_SDK.client_name,
         [core_data_model],
         CORE_SDK.instance_space,
+    )
+
+
+@pytest.fixture(scope="session")
+def turbine_multi_api_generator(turbine_data_model: dm.DataModel[dm.View]) -> MultiAPIGenerator:
+    return MultiAPIGenerator(
+        WIND_TURBINE.top_level_package,
+        WIND_TURBINE.client_name,
+        [turbine_data_model],
+        WIND_TURBINE.instance_space,
     )